fixup! feat: configure main window size
- Use Double for main window width/height to match the settings contract.

- Avoid recursive didSet by storing through a private backing property with a non-recursing computed setter.

+++ Macterm/Settings/SettingsView.swift
@@ -250,9 +250,9 @@ private struct AppearanceSettings: View {
     @State
     private var paneDimOpacity: Double = Preferences.shared.paneDimOpacity
     @State
-    private var mainWindowWidth: Int = Preferences.shared.mainWindowWidth
+    private var mainWindowWidth: Double = Preferences.shared.mainWindowWidth
     @State
-    private var mainWindowHeight: Int = Preferences.shared.mainWindowHeight
+    private var mainWindowHeight: Double = Preferences.shared.mainWindowHeight
 
     var body: some View {
         Form {
